Events and ToDos can be created now! Check the "New Event" tab. Fixed a bug where you couldn't delete events that hadn't been saved yet. Theoretically the app is functionable as a planner!

+++ Saves/DataManager.py
@@ -111,7 +111,8 @@ def delete_object(object : Note | Notebook | RepeatingEvent | Event | ToDo):
     fileName += '.json'
     
     filePath = savePath / saveFolderName / fileName
-    filePath.unlink()
+    if filePath.exists():
+        filePath.unlink()
